+++ docs/CHANGELOG.md
@@ -37,6 +37,7 @@
 - Correction d'un problème lors de la suppression des listes de case
 - Correction d'un problème d'affichage lors de l'export en PDF dans certains cas
 - L'action "Sélectionner une liste choisie" est maintenant simulée en mode configuration
+- Les images des configurations par défaut sont maintenant directement affichées
 
 
 ## 1.2.1 - 12/12/2021